+++ bakedanuki/bakedanuki-util/python/bd_util/ui/binding/float/view/__init__.py
@@ -5,6 +5,7 @@ from .slider import FloatSlider
 from .slider_spin_box import FloatSliderSpinBox
 from .range_slider_spin_box import FloatRangeSliderSpinBox
 from .step_spin_box import FloatStepMode, FloatStepSpinBox
+from .value_step_spin_box import FloatValueStepSpinBox
 
 __all__ = [
     "FloatSpinBox",
@@ -14,4 +15,5 @@ __all__ = [
     "FloatRangeSliderSpinBox",
     "FloatStepMode",
     "FloatStepSpinBox",
+    "FloatValueStepSpinBox",
 ]